knnnew (Heroku)/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jws==0.1.3
MarkupSafe==1.1.1
numpy==1.19.1
oauth2client==4.1.2
pandas==1.1.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
Pyrebase4==4.3.0
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==1.0.1
